Remove oncecell from integration tests.
