+++ crates/test_integration/src/tests/mod.rs
@@ -1,1 +1,1 @@
-pub mod api_integration;
+pub mod flow;
